+++ back-end/jongmaihub/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 from .models import *
 
+
 admin.site.register(ServiceProvider)
 admin.site.register(Service)
 admin.site.register(Staff)
