+++ back-end/jongmaihub/serializers/appointment_serializer.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+from ..models import Appointment
+
+
+class AppointmentSerializer(serializers.ModelSerializer):
+    """Serializer for Appointment model."""
+    class Meta:
+        fields = "__all__"
+        model = Appointment
