+++ back-end/jongmaihub/serializers/customer_serializer.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+from ..models import Customer
+
+
+class CustomerSerializer(serializers.ModelSerializer):
+    """Serializer for Customer model."""
+    class Meta:
+        fields = "__all__"
+        model = Customer
