+++ back-end/jongmaihub/serializers/service_provider_serializer.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from ..models import ServiceProvider
+
+
+class ServiceProviderSerializer(serializers.ModelSerializer):
+    """Serializer for ServiceProvider model."""
+    class Meta:
+        fields = "__all__"
+        model = ServiceProvider
+
