+++ back-end/jongmaihub/serializers/service_serializer.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from ..models import Service
+
+
+class ServiceSerializer(serializers.ModelSerializer):
+    """Serializer for Service model."""
+    class Meta:
+        fields = "__all__"
+        model = Service
+
