+++ back-end/jongmaihub/views/__init__.py
@@ -1,0 +1,13 @@
+from .detail_appointment import DetailAppointment
+from .detail_customer import DetailCustomer
+from .detail_rating import DetailRating
+from .detail_service import DetailService
+from .detail_service_provider import DetailServiceProvider
+from .detail_staff import DetailStaff
+from .list_appointment import ListAppointment
+from .list_appointment_by_provider import ListAppointmentsByProvider
+from .list_customer import ListCustomer
+from .list_rating import ListRating
+from .list_service import ListService
+from .list_service_provider import ListServiceProvider
+from .list_staff import ListStaff
