+++ back-end/jongmaihub/views/list_service.py
@@ -1,0 +1,17 @@
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticated
+from ..renderers import EncryptorRenderer
+from ..models import Service
+from ..serializers import ServiceSerializer
+
+
+class ListService(generics.ListCreateAPIView):
+    """ListService displays a list of all Service."""
+    permission_classes = (IsAuthenticated,)
+    renderer_classes = [EncryptorRenderer]
+    queryset = Service.objects.all()
+    serializer_class = ServiceSerializer
+    filter_backends = [DjangoFilterBackend]
+    filterset_fields = "__all__"
+
